+++ test/components/label_component_test.rb
@@ -11,16 +11,17 @@ class PrimerLabelComponentTest < Minitest::Test
     assert_text("private")
   end
 
-  def test_defaults_to_bg_blue
-    render_inline(Primer::LabelComponent.new(title: "foo")) { "private" }
-
-    assert_selector(".bg-blue")
-  end
-
   def test_supports_functional_schemes
     render_inline(Primer::LabelComponent.new(title: "foo", scheme: :danger)) { "private" }
 
     assert_selector(".Label--danger")
+  end
+
+  def test_deprecated_schemes
+    Primer::LabelComponent::DEPRECATED_SCHEME_OPTIONS.each do |scheme|
+      ActiveSupport::Deprecation.expects(:warn).with("#{scheme} is deprecated and will be removed in a future version.").once
+      render_inline(Primer::LabelComponent.new(title: "foo", scheme: scheme)) { "scheme" }
+    end
   end
 
   def test_falls_back_when_scheme_isn_t_valid