+++ CHANGELOG.md
@@ -2,14 +2,25 @@
 
 ## main
 
-* Update `LinkComponent` to use `Link--muted` instead of `muted-link`.
+* Update `CounterComponent` to accept functional schemes `primary` and `secondary`. Deprecate `gray` and `light_gray` schemes.
 
     *Manuel Puyol*
 
-* Promote `DetailsComponent`, `TextComponent`, `TimelineItemComponent`, and
+* Add `force_functional_colors` option to convert colors to functional. This change includes a deprecation warning in non-production environments that warns about non functional color usage.
+
+    *Manuel Puyol*
+
+* Promote `DetailsComponent`, `HeadingComponent`, `TextComponent`, `TimelineItemComponent`, and
   `PopoverComponent` to beta status.
 
     *Simon Taranto*
+
+* Update `LinkComponent`:
+  * use `Link--muted` instead of `muted-link`.
+  * accept `variant` and `underline` options.
+  * accept `:span` as a tag.
+
+    *Manuel Puyol*
 
 * Add `AutoComplete` and `AutoCompleteItem` components.